Change float-int benchmarks to use BigInteger
Reference Java methods in PyFloatBinary benchmarks, that mix int with
double, are changed to represent the int as a BigInteger. It seems a
fairer comparator if conversion is from that type.

Note the update to Java 11.0.9 and recompute all the benchmark tables.

+++ dy2bm/src/main/java/uk/co/farowl/vsj2dybm/evo4/PyFloatBinary.java
@@ -1,5 +1,6 @@
 package uk.co.farowl.vsj2dybm.evo4;
 
+import java.math.BigInteger;
 import java.util.concurrent.TimeUnit;
 
 import org.openjdk.jmh.annotations.Benchmark;
@@ -35,8 +36,8 @@ import uk.co.farowl.vsj2dy.evo4.AbstractProxy;
 @State(Scope.Thread)
 public class PyFloatBinary {
 
-    int iv = 6, iw = 7;
-    double v = 1.01 * iv, w = 1.01 * iw;
+    BigInteger iv = BigInteger.valueOf(6), iw = BigInteger.valueOf(7);
+    double v = 1.01 * iv.doubleValue(), w = 1.01 * iw.doubleValue();
     PyObject fvo = Py.val(v), fwo = Py.val(w);
     PyObject ivo = Py.val(iv), iwo = Py.val(iv);
 
@@ -52,7 +53,7 @@ public class PyFloatBinary {
     }
 
     @Benchmark
-    public double add_float_int_java() { return v + iw; }
+    public double add_float_int_java() { return v + iw.doubleValue(); }
 
     @Benchmark
     public PyObject add_float_int() throws Throwable {
@@ -60,7 +61,7 @@ public class PyFloatBinary {
     }
 
     @Benchmark
-    public double add_int_float_java() { return iv + w; }
+    public double add_int_float_java() { return iv.doubleValue() + w; }
 
     @Benchmark
     public PyObject add_int_float() throws Throwable {
